pyproject.toml
[project]
name = "face-hnfnu"
version = "0.1"
description = "simple face detector"
authors = []
dependencies = [
    "bcolz-zipline>=1.2.6",
    "blinker>=1.6.3",
    "certifi>=2023.7.22",
    "charset-normalizer>=3.3.0",
    "click>=8.1.7",
    "contourpy>=1.1.1",
    "cycler>=0.12.1",
    "filelock>=3.12.4",
    "Flask>=3.0.0",
    "fonttools>=4.43.1",
    "fsspec>=2023.9.2",
    "gevent>=23.9.1",
    "graphviz>=0.8.4",
    "greenlet>=3.0.0",
    "gunicorn>=21.2.0",
    "idna>=3.4",
    "importlib-metadata>=6.8.0",
    "importlib-resources>=6.1.0",
    "imutils>=0.5.4",
    "itsdangerous>=2.1.2",
    "Jinja2>=3.1.2",
    "kiwisolver>=1.4.5",
    "MarkupSafe>=2.1.3",
    "matplotlib>=3.7.3",
    "menpo>=0.11.0",
    "mpmath>=1.3.0",
    "mxnet>=1.7.0.post2",
    "networkx>=3.1",
    "numpy>=1.24.4",
    "opencv-python-headless>=4.7.0.72",
    "packaging>=23.2",
    "Pillow>=10.0.1",
    "prettytable>=3.9.0",
    "pyparsing>=3.1.1",
    "python-dateutil>=2.8.2",
    "PyYAML>=6.0.1",
    "requests>=2.31.0",
    "scipy>=1.10.1",
    "six>=1.16.0",
    "sympy>=1.12",
    "torch>=2.1.0",
    "tqdm>=4.66.1",
    "triton>=2.1.0",
    "typing-extensions>=4.8.0",
    "urllib3>=2.0.6",
    "wcwidth>=0.2.8",
    "Werkzeug>=3.0.0",
    "zipp>=3.17.0",
    "zope.event>=5.0",
    "zope.interface>=6.1",
    "faiss-cpu>=1.8.0",
    "packaging"
]
requires-python = ">=3.11, <3.12"
license = {text = "MIT"}

[build-system]
requires = ["pdm-pep517>=1.0.0"]
build-backend = "pdm.pep517.api"


[tool]
[tool.pdm]
